requirements.txt
anthropic>=0.40.0
Pillow>=10.0.0
pynput>=1.7.6
